@@ -1,47 +1,19 @@
 {
-    "name": "lzv-match-maker-ui",
-    "version": "0.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "lint": "eslint 'src/**/*.ts'",
-        "watch": "tsc -w",
-        "serve": "parcel serve ./src/index.html",
-        "debug": "run-p watch serve"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/thomaux/lzv-match-maker-ui.git"
-    },
-    "author": "",
-    "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/thomaux/lzv-match-maker-ui/issues"
-    },
-    "homepage": "https://github.com/thomaux/lzv-match-maker-ui#readme",
-    "private": true,
-    "dependencies": {
-        "inversify": "^5.0.1",
-        "inversify-inject-decorators": "^3.1.0",
-        "material-design-icons-iconfont": "^6.1.0",
-        "reflect-metadata": "^0.1.13",
-        "vue": "^2.6.12",
-        "vue-property-decorator": "^8.5.1",
-        "vue-router": "^3.4.3",
-        "vuelidate": "^0.7.5",
-        "vuetify": "^2.3.10"
-    },
-    "devDependencies": {
-        "@typescript-eslint/eslint-plugin": "^2.34.0",
-        "@typescript-eslint/parser": "^2.34.0",
-        "eslint": "^6.8.0",
-        "eslint-plugin-vue": "^6.2.2",
-        "npm-run-all": "^4.1.5",
-        "parcel-bundler": "^1.12.4",
-        "sass": "^1.26.10",
-        "typescript": "^3.9.7"
-    },
-    "alias": {
-        "vue": "./node_modules/vue/dist/vue.common.dev.js"
-    }
+  "name": "lzv-match-maker-ui",
+  "version": "0.0.0",
+  "scripts": {
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "vue": "^3.2.25",
+    "vue-auth0-plugin": "^2.4.4"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^2.0.0",
+    "typescript": "^4.4.4",
+    "vite": "^2.7.2",
+    "vue-tsc": "^0.29.8"
+  }
 }
